Corrected a small typo in the Route of the data controller.  Was missing a /
Also changed the inherited class of the controller from Controller to BaseController to mimic how it exists in the SampleSPA Data Controller.

With these two changes done the page now runs and hits a breakpoint in the GetSources endpoint.

Added additional knockout for displaying and editing the list of sources.
Added additional api controllers for saving and deleting.

todo: add the delete mechanic to the views.